+++ src/main/java/com/adobe/codingchallenge/repository/user/UserRepository.java
@@ -10,6 +10,6 @@ public interface UserRepository extends CrudRepository<User, Long> {
     //User findByUserIdAAndInactvDtIsNull(Integer userId);
 
     List<User> findUserByUserEmail(String email);
-    User findUserByUserId(int userId);
+    User findUserByUserId(long userId);
 
 }
